# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Park Sense Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  userId: number;
  slotId: number;
  lotId: number;
  lotName: string;
  slotNumber: string;
  status: BookingStatus;
  startTime: Date;
  endTime: Date;
  totalAmount: number;
  qrCode?: string | null;
  vehicleNumber?: string | null;
  createdAt: Date;
}