# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Park Sense Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  slotId: number;
  startTime: Date;
  endTime: Date;
  vehicleNumber?: string | null;
}