# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Park Sense Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewRequest {
  lotId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
}